\name{Choir}
\alias{Choir}
\docType{data}
\title{Choir}
\description{
Choir is the name of a dataset containing data of one performance by four singers. The performance has been recorded with 4 microphones plus one omni-microphone. The data are inter-onset-intervals extracted from either the 4 microphones and then merged, or the omni-microphone.}
\usage{data("Choir")}
\format{
  A list with the following entries:
  \itemize{
    \item{\code{Choir4MicOnsets}, a numeric array with inter-onset-intervals in milliseconds}
    \item{\code{ChoirOmniMicOnsets}, a numeric array with inter-onset-intervals in milliseconds}
}}
%\details{}
\references{
Leman, M.(2020).
Co-regulated timing in music ensembles - a Bayesian listener perspective. \emph{Submitted}.
}
\keyword{datasets}
